[GB.GTK] * NEW: SvgImage.Width and SvgImage.Height are now Float properties.
[GB.QT4]
* NEW: The SvgImage class has been implemented.


git-svn-id: svn://localhost/gambas/trunk@2584 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/lib/draw/gb.paint.h
@@ -109,8 +109,8 @@ typedef
 		struct GB_PAINT_DESC *desc;        // drawing driver
 		struct GB_PAINT *previous;         // previous drawing context
 		void *device;                      // drawing object
-		int width;                         // device width in device coordinates
-		int height;                        // device height in device coordinates
+		double width;                      // device width in device coordinates
+		double height;                     // device height in device coordinates
 		int resolutionX;                   // device horizontal resolution in DPI
 		int resolutionY;                   // device vertical resolution in DPI
 		PAINT_BRUSH *brush;                // current brush
